+++ lib/samples/impl/get_server/get_server_main.dart
@@ -1,17 +1,9 @@
-import 'package:get_cli/samples/interface/sample_interface.dart';
+// import 'package:get_cli/samples/interface/sample_interface.dart';
 
-class GetServerMainSample extends Sample {
-  GetServerMainSample() : super('lib/main.dart', overwrite: true);
+// class GetServerMainSample extends Sample {
+//   GetServerMainSample() : super('lib/main.dart', overwrite: true);
 
-  @override
-  Future<String> get content async =>
-      '''import 'package:get_server/get_server.dart';
-import 'app/routes/app_pages.dart';
-
-void main() {
-  runApp(GetServer(
-    getPages: AppPages.routes,
-  ));
-}
-  ''';
-}
+//   @override
+//   Future<String> get content async =>
+      
+// }