+++ lib/samples/interface/sample_interface.dart
@@ -1,9 +1,11 @@
+import 'dart:async';
+
 import 'package:get_cli/functions/create/create_single_file.dart';
 
 abstract class Sample {
   String path;
   bool overwrite;
-  Future<String> get content;
+  FutureOr<String> get content;
   Sample(this.path, {this.overwrite = false});
   Future<void> create() async {
     return await writeFile(path, await content, overwrite: overwrite);